Dockerfile: install bcc 0.4 on Fedora 26
To make gobpf buildable with bcc 0.4, we need to update the gobpf-ci
Docker image. Otherwise the CI fails because the new gobpf code based
on bcc 0.4 will not work with the original bcc 0.3 in gobpf-ci.

Since the iovisor repo doesn't provide rpm packages for bcc 0.4 yet,
we need to build bcc on our own.